Fixed a bug in the line drawing algorithm port caused by unsigned numbers being used in place where signed numbers were expected, leading to incorrect results and potential infinite loops.

+++ ZSharp/ZDrawing.h
@@ -1,6 +1,7 @@
 ﻿#ifndef ZDRAWING_H
 #define ZDRAWING_H
 
+#include <cstdint>
 #include <cstddef>
 
 #include "Framebuffer.h"
@@ -8,7 +9,7 @@
 
 namespace ZSharp {
 
-void DrawRunSlice(Framebuffer& framebuffer, std::size_t x1, std::size_t y1, std::size_t x2, std::size_t y2, ZColor color);
+void DrawRunSlice(Framebuffer& framebuffer, std::int32_t x1, std::int32_t y1, std::int32_t x2, std::int32_t y2, ZColor color);
 
 }
